// interfaces/restifyOpts.ts
/**
 * restifyOpts
 */

"use strict";


/* Node modules */


/* Third-party modules */


/* Files */


export interface IRestifyOpts {
    certificate?: string;
    formatters?: Object;
    handleUpgrades?: boolean;
    httpsServerOptions?: Object;
    key?: string;
    log?: Object;
    name?: string;
    spdy?: Object;
    version?: string | string[];
}